Fix for multiple ports
The interrupt handling has been revised in this commit to adopt
to the somewhat peculiar behavior of the SDK interrupt management.
Previously there was WDT resets whenever multiple SoftwareSerial
objects where used. This was due to the SDK calling the same
callback until all GPIO interrupts were reset.
Now only one interrupt handler is used for all objects and all
interrupts are reset here to avoid a loop.
Please note that this also means that the standard Arduino
attachInterrupt can not be used in combination with this library.