Make sketch filename case match folder name (#96)
The Arduino IDE requires the sketch folder name to match the filename of the primary sketch file. On a filename case-sensitive OS like Linux, this includes matching case. This change causes the example sketch to be accessible via the Arduino IDE's File > Examples > LIBRARYNAME menu after the library is installed.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification#library-examples